backend/requirements.txt
# Backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pdfplumber==0.11.4
pypdf==5.1.0
pydantic==2.10.0
pydantic-settings==2.6.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
psycopg[binary]>=3.1.0

# OCR dependencies
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
